login and signup changes from gokul
