+++ js/login.js
@@ -1,0 +1,38 @@
+import { initializeApp } from "https://www.gstatic.com/firebasejs/10.11.1/firebase-app.js";
+ import {getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword} from "https://www.gstatic.com/firebasejs/10.11.1/firebase-auth.js";
+ import{getFirestore, setDoc, doc} from "https://www.gstatic.com/firebasejs/10.11.1/firebase-firestore.js"
+ 
+ const firebaseConfig = {
+    apiKey: "AIzaSyAIuCN5NMapt-HyvTWmEXPOhXTdBYlVhOk",
+    authDomain: "login-backend-59af8.firebaseapp.com",
+    projectId: "login-backend-59af8",
+    storageBucket: "login-backend-59af8.appspot.com",
+    messagingSenderId: "1097943042656",
+    appId: "1:1097943042656:web:174161d4af1d7017cad105",
+    measurementId: "G-Q9C3E9L2P6"
+ };
+
+ const signIn=document.getElementById('submitSignIn');
+ signIn.addEventListener('click', (event)=>{
+    event.preventDefault();
+    const email=document.getElementById('email').value;
+    const password=document.getElementById('password').value;
+    const auth=getAuth();
+
+    signInWithEmailAndPassword(auth, email,password)
+    .then((userCredential)=>{
+        showMessage('login is successful', 'signInMessage');
+        const user=userCredential.user;
+        localStorage.setItem('loggedInUserId', user.uid);
+        window.location.href='home.html';               //********************page that we want to go -Aftab*******************/
+    })
+    .catch((error)=>{
+        const errorCode=error.code;
+        if(errorCode==='auth/invalid-credential'){
+            showMessage('Incorrect Email or Password', 'signInMessage');
+        }
+        else{
+            showMessage('Account does not Exist', 'signInMessage');
+        }
+    })
+ })
