+++ login.html
@@ -1,53 +1,102 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Login - Team Generator</title>
-    <link rel="stylesheet" href="css/login.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-</head>
-<body>
-    
-    <div class="page d-flex justify-content-center">
-        <div class="container bg-light">
-          <div class="row main-row">
-            <div class="col-md-6 d-flex main-svg">
-              <img class="svg-img" src="assets/images/Book-reading.svg" alt="Image" class="img-fluid">
-            </div>
-            <div class="col-md-6 d-flex main-content">
-              <div class="row">
-                <div class="col-md-11">
-                  <div class="mb-4">
-                  <h2 class="display-3 text-center">Login <span style="color: #AD50A7;">Here</span></h2>
-                  <p class="mb-4 d-flex gap-2">Don't have an account ? <br><a class="signup-redirect" href="signup.html" style="font-weight: bold;">Sign Up</a>.</p>
+    <link rel="stylesheet" href="css/login.css" />
+    <script type="module" src="js/login.js"></script>
+ 
+    <style>
+      .messageDiv {
+        background-color: hsl(327, 90%, 28%);
+        color: white;
+        padding: 10px 20px;
+        margin: 10px;
+        border-radius: 5px;
+        font-size: 1rem;
+        opacity: 0;
+        animation: fadeOut 7s forwards;
+      }
+    </style>
+ 
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
+      crossorigin="anonymous"
+    />
+  </head>
+  <body>
+    <div class="page d-flex justify-content-center" id="signIn">
+      <div class="container bg-light">
+        <div class="row main-row">
+          <div class="col-md-6 d-flex main-svg">
+            <img
+              class="svg-img"
+              src="assets/images/Book-reading.svg"
+              alt="Image"
+              class="img-fluid"
+            />
+          </div>
+          <div class="col-md-6 d-flex main-content">
+            <div class="row">
+              <div class="col-md-11">
+                <div class="mb-4">
+                  <h2 class="display-3 text-center">
+                    Login <span style="color: #ad50a7">Here</span>
+                  </h2>
+                  <p class="mb-4 d-flex gap-3">
+                    Don't have an account ? <br /><a
+                      href="signup.html"
+                      style="font-weight: bold"
+                      >Sign Up</a
+                    >.
+                  </p>
+                  <div
+                    id="signInMessage"
+                    class="messageDiv"
+                    style="display: none"
+                  ></div>
                 </div>
                 <form action="#" method="post">
                   <div class="form-group first mb-4">
                     <label for="username">E-mail</label>
-                    <input type="email" class="form-control opacity-50" id="email" placeholder="Enter your email">
-    
+                    <input
+                      type="email"
+                      class="form-control opacity-50"
+                      id="email"
+                      placeholder="Enter your email"
+                    />
                   </div>
                   <div class="form-group last mb-4">
                     <label for="password">Password</label>
-                    <input type="password" class="form-control opacity-50" id="password" placeholder="Enter your password">
-                    
+                    <input
+                      type="password"
+                      class="form-control opacity-50"
+                      id="password"
+                      placeholder="Enter your password"
+                    />
                   </div>
-    
-                  <input type="submit" id="submit" value="Login" class="btnform fw-bold p-2 ps-4 pe-4 btn-block">
-    
+ 
+                  <input
+                    type="submit"
+                    id="submitSignIn"
+                    value="Login"
+                    class="btnform fw-bold p-2 ps-4 pe-4 btn-block"
+                  />
                 </form>
-                </div>
               </div>
-              
             </div>
           </div>
         </div>
-        </div>
-    
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-    <script src="js/login.js"></script>
-</body>
+      </div>
+    </div>
+ 
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+      crossorigin="anonymous"
+    ></script>
+  </body>
 </html>
-
