+++ signup.html
@@ -1,82 +1,114 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <style>
+      .messageDiv {
+        background-color: hsl(327, 90%, 28%);
+        color: white;
+        padding: 10px 20px;
+        margin: 10px;
+        border-radius: 5px;
+        font-size: 1rem;
+        opacity: 0;
+        animation: fadeOut 7s forwards;
+      }
+    </style>
     <title>Sign Up - Team Generator</title>
-    <link rel="stylesheet" href="css/signup.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-</head>
-<body>
-    <div class="page d-flex justify-content-center">
-        <div class="container bg-light">
-          <div class="row main-row">
-            <div class="col-md-6 d-flex main-svg">
-              <img class="svg-img" src="assets/images/3236267.jpg" alt="Image" class="img-fluid">
-            </div>
-            <div class="col-md-6 d-flex main-content">
-              <div class="row">
-                <div class="col-md-11">
-                  <div class="mb-4">
-                  <h2 class="display-3 text-center">Register <span style="color: #AD50A7;">Now</span></h2>
-                  <p class="mb-4 fs-6 text-center"><a class="login-redirect" href="login.html">Login</a> if you already have an account</p>
+    <link rel="stylesheet" href="css/signup.css" />
+    <script type="module" src="js/signup.js"></script>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
+      crossorigin="anonymous"
+    />
+  </head>
+  <body>
+    <div class="page d-flex justify-content-center" id="signup">
+      <div class="container bg-light">
+        <div class="row main-row">
+          <div class="col-md-6 d-flex main-svg">
+            <img
+              class="svg-img"
+              src="assets/images/3236267.jpg"
+              alt="Image"
+              class="img-fluid"
+            />
+          </div>
+          <div class="col-md-6 d-flex main-content">
+            <div class="row">
+              <div class="col-md-11">
+                <div class="mb-4">
+                  <h2 class="display-3 text-center">
+                    Register <span style="color: #ad50a7">Now</span>
+                  </h2>
+                  <p class="mb-4 fs-6 text-center">
+                    <a class="login-redirect" href="login.html">Login</a> if you
+                    already have an account
+                  </p>
+                  <div
+                    id="signUpMessage"
+                    class="messageDiv"
+                    style="display: none"
+                  ></div>
                 </div>
                 <form action="#" method="post">
                   <div class="form-group first mb-4">
                     <label for="username">Username</label>
-                    <input type="text" class="form-control opacity-75" id="username" placeholder="Enter your username">
-    
+                    <input
+                      type="text"
+                      class="form-control opacity-75"
+                      id="username"
+                      placeholder="Enter your username"
+                    />
                   </div>
                   <div class="form-group second mb-4">
                     <label for="rPassword">Password</label>
-                    <input type="password" class="form-control opacity-75" id="rPassword" placeholder="Create a password">
-                    
+                    <input
+                      type="password"
+                      class="form-control opacity-75"
+                      id="rPassword"
+                      placeholder="Create a password"
+                    />
                   </div>
                   <div class="form-group third mb-4">
                     <label for="rEmail">Email</label>
-                    <input type="email" class="form-control opacity-75" id="rEmail" placeholder="Enter your e-mail">
-                    
+                    <input
+                      type="email"
+                      class="form-control opacity-75"
+                      id="rEmail"
+                      placeholder="Enter your e-mail"
+                    />
                   </div>
                   <div class="form-group last mb-4">
-                    <label for="phone">Phone</label>
-                    <input type="text" class="form-control opacity-75" id="phone" placeholder="Enter your phone number">
-                    
+                    <label for="phonenumber">Phone</label>
+                    <input
+                      type="text"
+                      class="form-control opacity-75"
+                      id="phonenumber"
+                      placeholder="Enter your Phone number"
+                    />
                   </div>
-                  
-                  <!-- <div class="d-flex mb-5 align-items-center">
-                    <label class="control control--checkbox mb-0"><span class="caption">Remember me</span>
-                      <input type="checkbox" checked="checked"/>
-                      <div class="control__indicator"></div>
-                    </label>
-                    <span class="ml-auto"><a href="#" class="forgot-pass">Forgot Password</a></span> 
-                  </div> -->
-    
-                  <input type="submit" value="Register" class="btnform p-2 pe-4 ps-4 fw-bold btn-block">
-    
-                  <!-- <span class="d-block text-center my-4 text-muted"> or Sign In with</span> -->
-                  
-                  <!-- <div class="social-login text-center">
-                    <a href="#" class="facebook">
-                      <span class="icon-facebook mr-3"></span> 
-                    </a>
-                    <a href="#" class="twitter">
-                      <span class="icon-twitter mr-3"></span> 
-                    </a>
-                    <a href="#" class="google">
-                      <span class="icon-google mr-3"></span> 
-                    </a>
-                  </div> -->
+                  <input
+                    type="submit"
+                    value="Submit"
+                    id="submitSignUp"
+                    class="btnform p-2 pe-4 ps-4 fw-bold btn-block"
+                  />
                 </form>
-                </div>
               </div>
-              
             </div>
           </div>
         </div>
-        </div>
-    
-    
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-    <script src="js/signup.js"></script>
-</body>
+      </div>
+    </div>
+ 
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+      crossorigin="anonymous"
+    ></script>
+  </body>
 </html>
